Update 'All your cluster are belong to us' icon
We we're using the icon user-secret for two talks (this one and 'eBPF,
the secret weapon of linux'). To let every talk have a distinctive icon
I've chosen a new one for Jorge's talk.

I choose 'fa-space-shuttle' as a reference for 'Zero Wing', the
videogame where the meme AYBABTU comes from.